bugfix: type cast must happen after comparison
Language reference:

> Casting from a larger integer to a smaller integer (e.g. u32 -> u8) will truncate

https://doc.rust-lang.org/reference/expressions/operator-expr.html#type-cast-expressions

The above means only the lower bits are used for the comparison, if the
length of the name was `257`, the cast would convert it to `1` and the
value would be wrongly considered inside the range.

+++ assembly/src/errors.rs
@@ -525,7 +525,7 @@ pub enum LabelError {
     EmptyLabel,
     InvalidFirstLetter(String),
     InvalidChars(String),
-    LabelTooLong(String, u8),
+    LabelTooLong(String, usize),
     Uppercase(String),
 }
 
@@ -542,7 +542,7 @@ impl LabelError {
         Self::InvalidFirstLetter(label.to_string())
     }
 
-    pub fn label_too_long(label: &str, max_len: u8) -> Self {
+    pub fn label_too_long(label: &str, max_len: usize) -> Self {
         Self::LabelTooLong(label.to_string(), max_len)
     }
